some tweaks
