+++ flexget/tests/cassettes/test_telegram.TestTelegramNotifier.test_send_image_as_document
@@ -33,7 +33,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Wed, 19 Feb 2025 01:42:58 GMT
+      - Wed, 19 Feb 2025 02:27:22 GMT
       Server:
       - nginx/1.18.0
       Strict-Transport-Security:
@@ -75,7 +75,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Wed, 19 Feb 2025 01:42:58 GMT
+      - Wed, 19 Feb 2025 02:27:23 GMT
       Server:
       - nginx/1.18.0
       Strict-Transport-Security:
@@ -104,7 +104,7 @@ interactions:
     uri: https://api.telegram.org/bot7617087239:AAGUy118YHbBvGNwkDo4CDehF4gFgXq2ZqE/sendMessage
   response:
     body:
-      string: '{"ok":true,"result":{"message_id":28,"from":{"id":7617087239,"is_bot":true,"first_name":"lk","username":"dddxgggfx_bot"},"chat":{"id":1394032416,"first_name":"o","last_name":"0","username":"chehxlpp","type":"private"},"date":1739929379,"text":"message"}}'
+      string: '{"ok":true,"result":{"message_id":35,"from":{"id":7617087239,"is_bot":true,"first_name":"lk","username":"dddxgggfx_bot"},"chat":{"id":1394032416,"first_name":"o","last_name":"0","username":"chehxlpp","type":"private"},"date":1739932043,"text":"message"}}'
     headers:
       Access-Control-Allow-Methods:
       - GET, POST, OPTIONS
@@ -119,7 +119,68 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Wed, 19 Feb 2025 01:42:59 GMT
+      - Wed, 19 Feb 2025 02:27:23 GMT
+      Server:
+      - nginx/1.18.0
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+    status:
+      code: 200
+      message: OK
+- request:
+    body: !!binary |
+      LS02OTM0NjQzOWRmZWI0MTUyMWJkZDM2NDMwOGI1OTM4Zg0KQ29udGVudC1EaXNwb3NpdGlvbjog
+      Zm9ybS1kYXRhOyBuYW1lPSJjaGF0X2lkIg0KDQoxMzk0MDMyNDE2DQotLTY5MzQ2NDM5ZGZlYjQx
+      NTIxYmRkMzY0MzA4YjU5MzhmDQpDb250ZW50LURpc3Bvc2l0aW9uOiBmb3JtLWRhdGE7IG5hbWU9
+      InBob3RvIjsgZmlsZW5hbWU9ImRvY3VtZW50LmpwZyINCkNvbnRlbnQtVHlwZTogaW1hZ2UvanBl
+      Zw0KDQr/2P/gABBKRklGAAEBAQB4AHgAAP/bAEMAAgEBAgEBAgICAgICAgIDBQMDAwMDBgQEAwUH
+      BgcHBwYHBwgJCwkICAoIBwcKDQoKCwwMDAwHCQ4PDQwOCwwMDP/bAEMBAgICAwMDBgMDBgwIBwgM
+      DAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDP/AABEIAHEA
+      DgMBIgACEQEDEQH/xAAfAAABBQEBAQEBAQAAAAAAAAAAAQIDBAUGBwgJCgv/xAC1EAACAQMDAgQD
+      BQUEBAAAAX0BAgMABBEFEiExQQYTUWEHInEUMoGRoQgjQrHBFVLR8CQzYnKCCQoWFxgZGiUmJygp
+      KjQ1Njc4OTpDREVGR0hJSlNUVVZXWFlaY2RlZmdoaWpzdHV2d3h5eoOEhYaHiImKkpOUlZaXmJma
+      oqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4eLj5OXm5+jp6vHy8/T19vf4+fr/
+      xAAfAQADAQEBAQEBAQEBAAAAAAAAAQIDBAUGBwgJCgv/xAC1EQACAQIEBAMEBwUEBAABAncAAQID
+      EQQFITEGEkFRB2FxEyIygQgUQpGhscEJIzNS8BVictEKFiQ04SXxFxgZGiYnKCkqNTY3ODk6Q0RF
+      RkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqCg4SFhoeIiYqSk5SVlpeYmZqio6Slpqeoqaqy
+      s7S1tre4ubrCw8TFxsfIycrS09TV1tfY2dri4+Tl5ufo6ery8/T19vf4+fr/2gAMAwEAAhEDEQA/
+      AP38ooooAKKKKACiiigAooooAKKKKACiiigAooooAKKKKAP/2Q0KLS02OTM0NjQzOWRmZWI0MTUy
+      MWJkZDM2NDMwOGI1OTM4Zi0tDQo=
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate, br, zstd
+      connection:
+      - keep-alive
+      content-length:
+      - '932'
+      content-type:
+      - multipart/form-data; boundary=69346439dfeb41521bdd364308b5938f
+      host:
+      - api.telegram.org
+      user-agent:
+      - python-telegram-bot v21.10 (https://python-telegram-bot.org)
+    method: POST
+    uri: https://api.telegram.org/bot7617087239:AAGUy118YHbBvGNwkDo4CDehF4gFgXq2ZqE/sendPhoto
+  response:
+    body:
+      string: '{"ok":true,"result":{"message_id":36,"from":{"id":7617087239,"is_bot":true,"first_name":"lk","username":"dddxgggfx_bot"},"chat":{"id":1394032416,"first_name":"o","last_name":"0","username":"chehxlpp","type":"private"},"date":1739932044,"photo":[{"file_id":"AgACAgUAAxkDAAMkZ7VBjFGdEzAvhk197gw0_TGvxowAArzAMRu1RLBVeotFmc2WGnMBAAMCAANzAAM2BA","file_unique_id":"AQADvMAxG7VEsFV4","file_size":292,"width":11,"height":90},{"file_id":"AgACAgUAAxkDAAMkZ7VBjFGdEzAvhk197gw0_TGvxowAArzAMRu1RLBVeotFmc2WGnMBAAMCAANtAAM2BA","file_unique_id":"AQADvMAxG7VEsFVy","file_size":296,"width":14,"height":113}]}}'
+    headers:
+      Access-Control-Allow-Methods:
+      - GET, POST, OPTIONS
+      Access-Control-Allow-Origin:
+      - '*'
+      Access-Control-Expose-Headers:
+      - Content-Length,Content-Type,Date,Server,Connection
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '592'
+      Content-Type:
+      - application/json
+      Date:
+      - Wed, 19 Feb 2025 02:27:24 GMT
       Server:
       - nginx/1.18.0
       Strict-Transport-Security:
