+++ flexget/tests/test_archives.py
@@ -1,3 +1,6 @@
+import pytest
+
+
 class TestArchiveFilter:
     config = """
         tasks:
@@ -11,6 +14,7 @@ class TestArchiveFilter:
                 archives: yes
     """
 
+    @pytest.mark.require_optional_deps
     def test_rar(self, execute_task):
         """Test RAR acceptance"""
         task = execute_task('test_archives')
