+++ flexget/tests/test_plex_watchlist.py
@@ -4,6 +4,7 @@ PLEX_USERNAME = 'flexget_flexget'
 PLEX_PASSWORD = 'flexget_flexget'
 
 
+@pytest.mark.require_optional_deps
 @pytest.mark.online
 class TestPlex:
     config = f"""
